+++ src/Core/SolutionName.Domain/Entities/Identity/RoleClaimEntity.cs
@@ -30,9 +30,9 @@ public class RoleClaimEntity : IdentityRoleClaim<int>, IAuditableEntity<int>
     public RoleEntity Role { get; set; }
 
     public DateTime CreatedDate { get; set; }
-    public long CreatedUserId { get; set; }
+    public int CreatedUserId { get; set; }
     public DateTime UpdatedDate { get; set; }
-    public long? UpdatedUserId { get; set; }
+    public int? UpdatedUserId { get; set; }
     public bool IsDeleted { get; set; }
     public int OrderIndex { get; set; }
