+++ src/Core/SolutionName.Domain/Entities/Identity/UserRoleEntity.cs
@@ -16,9 +16,9 @@ public class UserRoleEntity : IdentityUserRole<int>, IAuditableEntity<int>
     public int Id { get; set; }
 
     public DateTime CreatedDate { get; set; }
-    public long CreatedUserId { get; set; }
+    public int CreatedUserId { get; set; }
     public DateTime UpdatedDate { get; set; }
-    public long? UpdatedUserId { get; set; }
+    public int? UpdatedUserId { get; set; }
     public bool IsDeleted { get; set; }
     public int OrderIndex { get; set; }
